Add routing
Add routes to our 'app-routing.module.ts' file.
Create three components that will be use to show in the main page depending on the route we use.
Create a sidebar component that have a list with the name of our three components.
Configure the list elements to change the route when they are clicked using the 'routerLink' property.