Get a country or null when searching by id.
Modify the 'searchCountryByCode' method in the 'countries.service.ts' file. If it gets a 'Country' array, map the array and return the first element. If it not get any country, return 'null'.
Implement 'switchMap' from rxjs in the 'country-page.component.ts' to reduce code when using an 'Observable' inside another 'Observable'.
Use 'Router' to navigate to a specific route when no country is found.